# infrastructure/api.yaml
swagger: '2.0'
info:
  title: HWSH API
  description: HWSH API
  version: 1.0.0
host: api.hello-world.sh
x-google-endpoints:
- name: api.hello-world.sh
  allowCors: True
schemes:
  - https
produces:
  - application/json
securityDefinitions:
  firebase:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://securetoken.google.com/hwsh-api"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "hwsh-api"
paths:
  /static/{path}:
    get:
      summary: Get static assets
      operationId: get-static
      parameters:
        - in: path
          name: path
          type: string
          required: true
          description: path of static asset
      x-google-backend:
        address: https://hwsh-api-static-e37jkmtdgq-ez.a.run.app 
        protocol: h2
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        '200':
          description: A successful response
          schema:
            type: string
  /machine:
    get:
      summary: Describe the machine state
      security:
        - firebase: []
      operationId: get-machine
      x-google-backend:
        address: https://hwsh-api-machine-e37jkmtdgq-ez.a.run.app
        protocol: h2
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        '200':
          description: A successful response
          schema:
            type: string
    patch:
      summary: Switch on machine and connect
      security:
        - firebase: []
      operationId: patch-machine
      x-google-backend:
        address: https://hwsh-api-machine-e37jkmtdgq-ez.a.run.app
        protocol: h2
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        '200':
          description: A successful response
          schema:
            type: string
    delete:
      summary: Stop the machine
      security:
        - firebase: []
      operationId: delete-machine
      x-google-backend:
        address: https://hwsh-api-machine-e37jkmtdgq-ez.a.run.app
        protocol: h2
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        '200':
          description: A successful response
          schema:
            type: string
  